% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDiffs.R
\name{AllDiffs}
\alias{AllDiffs}
\title{Deprecated}
\usage{
AllDiffs(Var1, Var2, Lengths = F)
}
\arguments{
\item{Var1}{A vector}

\item{Var2}{Another vector}

\item{Lengths}{A boolean to return the lengths}
}
\value{
\code{Lengths}
}
\description{
Currently, returns \code{FALSE} all the time.
}
\note{
This function does not make sense, and used nowhere.
}
