% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arrange_ggplot2.R
\name{vp.layout}
\alias{vp.layout}
\alias{arrange_ggplot2}
\title{Arrange ggplot2}
\usage{
vp.layout(x, y)

arrange_ggplot2(dots, nrow = NULL, ncol = NULL, as.table = FALSE)
}
\arguments{
\item{y}{}

\item{as.table}{}
}
\description{
Arrange ggplot2

Title
}
