% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsReml_Model_Plot.R, R/Asreml_Diagnostics.R,
%   R/Asreml_Repeatability.R
\name{ASRPlot}
\alias{ASRPlot}
\alias{AsDiagnostics}
\alias{AsRepPlot}
\title{AsReml Plot}
\usage{
ASRPlot(model)

AsDiagnostics(model)

AsRepPlot(model)
}
\arguments{
\item{model}{}

\item{model}{}

\item{model}{}
}
\description{
AsReml Plot

asreml diagnostics

AsReml Repeatability
}
