% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarBarGraph.R
\name{data_summary}
\alias{data_summary}
\alias{BarBarGraph}
\title{BarBarGraph}
\usage{
data_summary(data, varname, groupnames)

BarBarGraph(data, group, x, y)
}
\arguments{
\item{data}{}

\item{groupnames}{}

\item{y}{}
}
\description{
BarBarGraph
}
