% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_Animal_Counting.R,
%   R/Cumulative_Animal_Counting_by_days.R
\name{CumulativeAnimals}
\alias{CumulativeAnimals}
\title{Title}
\usage{
CumulativeAnimals(df)

CumulativeAnimals(df)
}
\arguments{
\item{df}{}

\item{unit}{}
}
\description{
Title

Cumulative Animals
}
\note{
There are two version of this function, that does two things.
}
