% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Effects_Plot.R
\name{Efxplot}
\alias{Efxplot}
\title{EfxPlotComp}
\usage{
Efxplot(ModelList, sig = TRUE, ModelNames = NULL, VarNames = NULL,
  VarOrder = NULL, tips = 0.2)
}
\arguments{
\item{tips}{}
}
\description{
EfxPlotComp
}
