% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_EvenCount.R
\name{Evencount}
\alias{Evencount}
\title{Evencount}
\usage{
Evencount(df)
}
\arguments{
\item{df}{A data.frame that contains \code{Samples} and \code{Day}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Evencount
}
