% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formula_generation.R
\name{FormulaGen}
\alias{FormulaGen}
\title{Generates all possible combinations of a set of covariates ####}
\usage{
FormulaGen(Response, covar)
}
\arguments{
\item{covar}{}
}
\description{
Generates all possible combinations of a set of covariates ####
}
