% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom_flat_violin.R
\docType{data}
\name{geom_flat_violin}
\alias{geom_flat_violin}
\alias{GeomFlatViolin}
\title{Title}
\usage{
geom_flat_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{datasets}
