% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_INLA_Field.R
\name{ggField}
\alias{ggField}
\title{ggplot for INLA fields}
\usage{
ggField(Model, Mesh, Groups = 1, Fill = "Discrete", Boundary = NULL,
  BoundaryWidth = 1)
}
\arguments{
\item{BoundaryWidth}{}
}
\description{
ggplot for INLA fields
}
