% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gretworks.R
\name{Gretworks}
\alias{Gretworks}
\title{Gretworks}
\usage{
Gretworks(Data, Columns = c("From", "To"), Mode = "Uni",
  Weighted = T, Directed = F, Scaled = F, Nodes = "All",
  Return = "Graph")
}
\arguments{
\item{Return}{}
}
\description{
Gretworks
}
