% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA_DIC_Plot.R, R/INLA_DIC_Retrieval.R,
%   R/INLA_Model_Addition.R, R/INLA_Model_Selection.R, R/INLA_Range_Plot.R,
%   R/INLA_Repeatability.R
\name{INLADICFig}
\alias{INLADICFig}
\alias{MDIC}
\alias{INLAModelAdd}
\alias{INLAModelSel}
\alias{INLARange}
\alias{INLARep}
\alias{INLARepPlot}
\title{Retrieving DIC from INLA models quickly ####}
\usage{
INLADICFig(ModelList, ModelNames = NULL, Just = F, CutOff = 2)

MDIC(model)

INLAModelAdd(Response, Explanatory, Add, Random = NULL,
  RandomModel = NULL, Family, Data, Delta = 2)

INLAModelSel(Response, Explanatory, Random = NULL, RandomModel = NULL,
  Family, Data, Delta = 2)

INLARange(ModelList, MaxRange, MeshList, ModelNames = NULL)

INLARep(Model, Family = "gaussian")

INLARepPlot(ModelList, ModelNames = NULL, VarNames = NULL,
  VarOrder = NULL, Just = F, DICOrder = F, CutOff = 0,
  Family = "gaussian", Residual = T, CI = F, Position = "stack",
  Plot = T)
}
\arguments{
\item{ModelNames}{}

\item{CutOff}{}

\item{model}{}

\item{Family}{}

\item{Delta}{}

\item{Plot}{}

\item{Delta}{}
}
\description{
Retrieving DIC from INLA models quickly ####

Deriving range of autocorrelation figures

INLA ICC function
}
