% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geographic_Region_Assignment.R
\name{LocToReg6}
\alias{LocToReg6}
\title{Jisca's Region Assignment}
\usage{
LocToReg6(E, N)

LocToReg6(E, N)
}
\arguments{
\item{N}{}

\item{N}{}
}
\description{
Jisca's Region Assignment

Title
}
\note{
There are two versions of this function.
}
\examples{
df$Reg6 <- with(df, LocToReg6(E, N))
ggplot(df, aes(E, N, colour = Reg6)) + geom_point() + coord_fixed()
}
