% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Largest_and_Smallest.R,
%   R/Maths_convenience_functions.R
\name{Largest}
\alias{Largest}
\alias{Smallest}
\alias{Prev}
\alias{logistic}
\alias{logit}
\title{Largest and smallest}
\usage{
Largest(vector, no = 5, fun = "raw")

Smallest(vector, no = 5)

Prev(x)

logistic(a)

logit(a)
}
\arguments{
\item{vector}{}

\item{no}{}

\item{fun}{}

\item{x}{}
}
\description{
Largest and smallest

Mathematical convenience functions
}
\note{
These function are represented as \code{*logis} functions.
}
