% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixPoints.R
\name{MatrixPoints}
\alias{MatrixPoints}
\title{Plot to compare matrix values}
\usage{
MatrixPoints(Matrix1, Matrix2, Names = NULL, Axes = NULL)
}
\arguments{
\item{Axes}{}
}
\description{
Plot to compare matrix values
}
