% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Factor_Comparison.R
\name{MCMCFactorComp}
\alias{MCMCFactorComp}
\title{Liverpool era factor comparison}
\usage{
MCMCFactorComp(model, vars)
}
\arguments{
\item{vars}{}
}
\description{
Liverpool era factor comparison
}
