% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Factor_Comparison_Multivariate.R
\name{MCMCFactorComp2}
\alias{MCMCFactorComp2}
\title{MCMC Factor Comparison for Multivariate Models}
\usage{
MCMCFactorComp2(Model, Response, Factor)
}
\arguments{
\item{Factor}{}
}
\description{
MCMC Factor Comparison for Multivariate Models
}
