% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Effect_Size_Comparison.R
\name{MCMCFxComp}
\alias{MCMCFxComp}
\title{Liverpool era effect size comparison}
\usage{
MCMCFxComp(model, vars, numeric = T)
}
\arguments{
\item{numeric}{}
}
\description{
Liverpool era effect size comparison
}
