% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeshInclude.R
\name{MeshInclude}
\alias{MeshInclude}
\title{MeshInclude}
\usage{
MeshInclude(xvar, yvar, xmin, xmax, ymin, ymax)
}
\arguments{
\item{ymax}{}
}
\description{
MeshInclude
}
\note{
This function does not seems to be done, therefore it is not exported.
}
