% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARows.R
\name{NARows}
\alias{NARows}
\title{Removing NA Rows}
\usage{
NARows(df, vars)
}
\arguments{
\item{vars}{}
}
\description{
Removing NA Rows
}
\note{
This function is effectively \code{\link[stats:na.omit]{stats::na.omit()}}, thus this is not
exported
}
