% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenCorr_Function.R
\name{PhenCorr}
\alias{PhenCorr}
\title{Phenotypic Correlation Calculation from a multivariate model}
\usage{
PhenCorr(model, Terms)
}
\arguments{
\item{Terms}{}
}
\description{
Phenotypic Correlation Calculation from a multivariate model
}
