% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting_Repeatability.R
\name{RepeatabilityReport}
\alias{RepeatabilityReport}
\title{Deriving Repeatability as a Function}
\usage{
RepeatabilityReport(Model)
}
\arguments{
\item{Model}{}
}
\value{
Currently, returns nothing.
}
\description{
Deriving Repeatability as a Function
}
