% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeSave.R
\name{SafeSave}
\alias{SafeSave}
\title{Quick workspace save}
\usage{
SafeSave()
}
\value{
Nothing
}
\description{
This saves to hard-drive. Use with care.
}
\note{
This function contains personal paths. Therefore it is not exported.
Use with care.
}
