% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nice_Scale_Function.R
\name{scale_this}
\alias{scale_this}
\title{Non-problematic scale function}
\usage{
scale_this(x)
}
\arguments{
\item{x}{}
}
\description{
Normalises a vector \code{x} such that mean is zero and variance is 1.
}
\details{
Missing values are ignored by default.
}
\seealso{
For a version that works on \code{data.frame} see \code{\link[=scale]{scale()}}.
}
