% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sina_Graph.R
\name{SinaGraph}
\alias{SinaGraph}
\title{SinaGraph}
\usage{
SinaGraph(df, x, y, z = x, Order = F, Just = F, Alpha = 1,
  Scale = "area", ColourGroups = F)
}
\arguments{
\item{ColourGroups}{}
}
\value{
A ggplot2 plot object
}
\description{
SinaGraph
}
