% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeChange.R
\name{ThemeChange}
\alias{ThemeChange}
\alias{ThemeRevert}
\title{Theme Changing}
\usage{
ThemeChange()

ThemeRevert()
}
\value{
Nothing
}
\description{
Needs fixing
}
\note{
This function should have an include tag for "Establishing Themes and Palettes.R"

This function should return to the theme that was present before this function was
invoked, i.e. using \code{CurrentTheme}.
}
