% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unwhich.R
\name{unwhich}
\alias{unwhich}
\title{Unwhich}
\usage{
unwhich(x, n)
}
\arguments{
\item{x}{result of \code{\link[=which]{which()}}}

\item{n}{size of the boolean mask to return}
}
\value{
A boolean mask representation of indices
}
\description{
Unwhich
}
\note{
\itemize{
\item An idea could be to default \code{n} to be the maximum.
\item Type-check that \code{x} is all integers
\item This is currently unexplored
}
}
\examples{
x <- rnorm(10)
unwhich(x >= 0, 10)

}
